Rename to arrow format utils
