+++ frontend/lib/src/components/widgets/DataFrame/columns/utils.ts
@@ -31,7 +31,7 @@ import "moment-timezone"
 import numbro from "numbro"
 import { sprintf } from "sprintf-js"
 
-import { formatPeriodType } from "@streamlit/lib/src/dataframes/arrowFormatting"
+import { formatPeriodType } from "@streamlit/lib/src/dataframes/arrowFormatUtils"
 import { Type as ArrowType } from "@streamlit/lib/src/dataframes/arrowTypeUtils"
 import { EmotionTheme } from "@streamlit/lib/src/theme"
 import {
